requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
fpdf
pyttsx3
